feat(data-scraper): fetch unified catalog.json instead of paging WS catalog
Replace the per-command catalog pagination and the ~2,555-call per-facility
WS detail fetches with a single HTTPS GET of /api/catalog.json (no auth/WS),
ETag-cached (.catalog.json/.catalog.etag) with If-None-Match and graceful
fallback to the cache on 304/429/network error. Split the unified bundle into
the legacy envelope files downstream importers and the kb generator consume:

- catalog_{ships,skills,recipes}.json
- catalog_items.json (items + modules combined, as the previous shape)
- catalog_modules.json (new; modules are now a distinct top-level section)
- catalog_facilities.json (single {"items":[...]} replacing facility_*.json
  and the facility_details/ directory)

Also fix a pre-existing recipe import break: the server made crafting_time
fractional, but the importer/knowledge struct/SQLite column were INTEGER.
Migrate RecipeDef.CraftingTime + RecipeJSON.CraftingTime to float64 and add
migration 45 rebuilding recipes.crafting_time as REAL (FK-safe cache table),
regenerating scripts/sql/initialize_database.sql.

Anchor the root-binary patterns in .gitignore: the unanchored "data-*" (and
siblings) matched cmd/data/data-scraper/ and other cmd/ source dirs, silently
ignoring new .go files there.

+++ cmd/data/import-catalog-recipes/main.go
@@ -20,7 +20,7 @@ type RecipeJSON struct {
 	Name            string                `json:"name"`
 	Description     string                `json:"description"`
 	Category        string                `json:"category"`
-	CraftingTime    int                   `json:"crafting_time"`
+	CraftingTime    float64               `json:"crafting_time"`
 	BaseQuality     int                   `json:"base_quality"`
 	SkillQualityMod int                   `json:"skill_quality_mod"`
 	RequiredSkills  map[string]int        `json:"required_skills"`
